ai-core/voice/stt_whisper.py
"""
Whisper Speech-to-Text Integration  
Actual implementation using OpenAI Whisper
"""

import logging
from pathlib import Path
from typing import Optional, Dict

logger = logging.getLogger(__name__)


class WhisperSTT:
    """
    Speech-to-Text using OpenAI Whisper
    Falls back gracefully if model not available
    """
    
    def __init__(self, model_size: str = "base", device: str = "cpu"):
        self.model_size = model_size
        self.device = device
        self.model = None
        self.use_mock = True
        
        try:
            import whisper
            logger.info(f"Loading Whisper model: {model_size}")
            self.model = whisper.load_model(model_size, device=device)
            self.use_mock = False
            logger.info("✓ Whisper model loaded")
        except ImportError:
            logger.warning("Whisper not installed, using mock transcription")
        except Exception as e:
            logger.warning(f"Whisper load failed: {e}, using mock")
    
    def transcribe_audio(self, audio_path: str, language: str = "en") -> dict:
        """Transcribe audio file to text"""
        
        if self.use_mock:
            return self._mock_transcribe(audio_path, language)
        
        try:
            result = self.model.transcribe(
                audio_path,
                language=language,
                fp16=(self.device == "cuda")
            )
            
            return {
                "text": result["text"].strip(),
                "language": result.get("language", language),
                "confidence": 0.95,
                "segments": result.get("segments", [])
            }
        except Exception as e:
            logger.error(f"Transcription failed: {e}")
            return self._mock_transcribe(audio_path, language)
    
    def _mock_transcribe(self, audio_path: str, language: str) -> dict:
        """Mock transcription for testing"""
        return {
            "text": "This is a mock transcription",
            "language": language,
            "confidence": 0.0,
            "is_mock": True
        }
    
    def transcribe_stream(self, audio_buffer: bytes, sample_rate: int = 16000) -> str:
        """Transcribe audio from byte buffer"""
        # TODO: Implement streaming support
        return "Streaming not yet implemented"
    
    def detect_language(self, audio_path: str) -> str:
        """Detect language of audio"""
        if self.use_mock:
            return "en"
        
        try:
            audio = self.model.transcribe(audio_path, language=None)
            return audio.get("language", "en")
        except:
            return "en"
    
    def is_loaded(self) -> bool:
        """Check if real model is loaded"""
        return not self.use_mock


class VoiceActivityDetector:
    """Simple voice activity detection"""
    
    def __init__(self, threshold: float = 0.5):
        self.threshold = threshold
    
    def is_speech(self, audio_buffer: bytes) -> bool:
        """Detect if audio contains speech"""
        # Simple energy-based detection
        return len(audio_buffer) > 1024
    
    def get_speech_segments(self, audio_path: str) -> list:
        """Extract speech segments from audio"""
        return [(0.0, 2.0)]


if __name__ == '__main__':
    logging.basicConfig(level=logging.INFO)
    
    stt = WhisperSTT(model_size="base")
    
    # Test with mock
    result = stt.transcribe_audio("test.wav")
    print(f"Transcription: {result['text']}")
    print(f"Using mock: {result.get('is_mock', False)}")
    print(f"Model loaded: {stt.is_loaded()}")
